requirements.txt
feedparser==6.0.11
beautifulsoup4==4.12.3
requests==2.31.0
pymongo==4.6.1
python-dotenv==1.0.1
curl-cffi==0.6.2
